+++ src/commands/NewCamelProjectCommand.ts
@@ -20,8 +20,10 @@ import { confirmDestructiveActionInSelectedFolder } from '../helpers/modals';
 
 export class NewCamelProjectCommand {
 	public static readonly ID_COMMAND_CAMEL_NEW_PROJECT = 'kaoto.camel.jbang.export';
+	public static readonly ID_COMMAND_CAMEL_NEW_PROJECT_FOLDER = 'kaoto.camel.jbang.export.folder';
+	public static readonly ID_COMMAND_CAMEL_NEW_PROJECT_WORKSPACE = 'kaoto.camel.jbang.export.workspace';
 
-	public async create(uri: Uri) {
+	public async create(uri: Uri, cwd: string) {
 		const runtime = await this.askForRuntime();
 		if (!runtime) {
 			return;
@@ -44,9 +46,8 @@ export class NewCamelProjectCommand {
 						return;
 					}
 				}
-				await new CamelExportJBangTask(currentWorkspace, uri.fsPath, input, runtime, outputFolder.fsPath).executeAndWaitWithProgress(
-					'Creating a new Camel project...',
-				);
+				const camelExportJBangTask = await CamelExportJBangTask.create(currentWorkspace, uri, input, runtime, outputFolder.fsPath, cwd);
+				await camelExportJBangTask.executeAndWaitWithProgress('Creating a new Camel project...');
 
 				// open the newly created project in a new vscode instance
 				await commands.executeCommand('vscode.openFolder', outputFolder, true);
@@ -127,6 +128,7 @@ export class NewCamelProjectCommand {
 			canSelectFiles: false,
 			openLabel: 'Select',
 			title: 'Select a folder to create the project in. ESC to cancel the project creation',
+			defaultUri: workspace.workspaceFolders?.[0]?.uri,
 		});
 		if (selectedFolders !== undefined) {
 			return selectedFolders[0];