+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - rework Integrations section of Kaoto view to bring tree-like structure
 - add buttons `Run: Folder` and `Run: Workspace` into Integrations section of Kaoto view
 - user setting allowing specify a list of local kamelet directories used for integrations run
+- add button `Export Folder` and `Export: Workspace` into Integrations section of Kaoto view
+- install Camel Kubernetes plugin into JBang by default
+- create Maven project with OpenShift deployment configuration
 
 # 2.8.0